Move danger to the test stage
Given GitLab Pipelines work based on pushes, and not MRs, it can happen
that the danger job starts before the MR is created. This leads to no
output from danger.

By moving it up one stage, a bit more MRs will be done on the first go
too.

An example of this happening is:
https://gitlab.com/gitlab-org/gitaly/-/jobs/64654110