Move Linguist to the Ruby sidecar
Currently, linguist is run by shelling out and spinning up a new Ruby
VM, require both Rugged and Linguist in there and extract the results.
This works, but had a couple of downsides, for one, the caching
mechanism wasn't available to be used. But also, some interesting hacks
were used to load the `languages.yaml` from linguist. But now they also
use json, which we missed an lead to:

1. https://gitlab.com/gitlab-org/gitaly/issues/1192
2. https://gitlab.com/gitlab-org/gitaly/issues/1042
3. https://gitlab.com/gitlab-org/gitaly/issues/1119

Downside of this approach is the fact that we're now doing more requests
on the sidecar, which might lead to ResourceExhausted errors, but
https://gitlab.com/gitlab-org/gitaly/merge_requests/700 is expected to
fix those.

Closes https://gitlab.com/gitlab-org/gitaly/issues/1192