Adds GRPC integration tests for client.Dial
client.Dial is the recommended way for Golang Gitaly clients to dial the
Gitaly service, given a connection string. The method takes a URL which
can refer to an insecure TCP connection, a secure TLS connection or an
(insecure) Unix socket connection. Additionally, it supports several
forms of URL for some of the URLs.

Unfortunately, we've found this interface to be fragile, having broken
during refactoring and GRPC library upgrades. This change is an
attempt to build some integration tests to catch these failures early,
rather than later on in the process.