Only return sanitized errors for Custom Hooks
pre-receive, post-receive and update custom hooks will now only return
error messages when a script exits with an error code if the error
messages have been prefixed in a particular way, and sent to stdout.

All messages sent to stderr during the script execution are now logged,
so sysadmins can see the full picture of the error.

This change has been made in order to re-enable the feature in GitLab
Rails of custom error messages being returned from custom hooks, which
had been disabled in Rails due to security concerns around this
potentially leaking things like stacktraces and other information that
is not intended to be public.

https://gitlab.com/gitlab-org/gitlab-ce/issues/48132