Avoid deadlocks in popen3
When there is a lot of stderr and/or stdout output from the file opened
by popen3, then the code can deadlock and hang forever.

See the note in the Ruby docs around the usage of popen3:

https://docs.ruby-lang.org/en/2.0.0/Open3.html#method-i-popen3

Also changing a use of open3 to capture3, as the code doesn't need to
write to standard in.