Custom hooks return stdout + stderr on success
Previously custom hooks would return the stdout of the executed script,
or if that was blank, then the stderr.

Now custom hooks will return the combined stdout and stderr messages.

Note, that the behaviour of these message only being returned when the
execute script exits with a success status code has been preserved.

A documentation change to Rails is in
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/25625

Issue https://gitlab.com/gitlab-org/gitlab-ce/issues/48132

Additionally, this brings it closer into line with the behaviour of
git-receive-pack, which we're trying to emulate.

From: https://git-scm.com/docs/githooks:

> Both standard output and standard error output are forwarded to git
> send-pack on the other end, so you can simply echo messages for the
> user.