Repo language detected no more that once a week
Noted by analyzing Sidekiq usage patterns[1], the commit languages
topped the charts in time spend. Most, almost all, of this time is spend
on Gitaly.

Now Gitaly could reject the call with a FailedPrecondition if it finds
it's scanning the repository too much. The heuristic to be used for this
would be cache modification time. The cache will be updated each
linguist run, and only RPC per repository is called at once from the
client side.

[1]: https://gitlab.com/gitlab-com/gl-infra/scalability/issues/64#note_254382111