hooks: Refactor passing through feature flags
Prior to this change, the value of the feature flag was passed through
multiple layers from the repository object all the way to the hooks. The
hooks however, have access to that repository already. This means
that the hooks could use this instance to request the feature flags.

Ideally there would be a bigger overarching scope that provides access
to the call. As this change now teaches the Repository instance about
feature flags too, which are inherently not their responsibility. For
now this is acceptable technical debt, is my judgement.