worktree: Add random suffix to worktree paths to obstruct path traversal
Currently, worktree paths used for operations like rebase or squash that
require access to checked-out files are generated deterministically by
using an operation-specific prefix as well as the ID of the given
operation. If given a path-traversal vulnerability, this makes it easy
to use those worktree paths as an attack vector to read arbitrary files
and directories.

To improve our defense-in-depth mechanisms, this commit adds a random
16-digit hex suffix so that the resulting path is not deterministic. As
we use blocks to manage worktrees, we know that the paths are not used
outside of a given block anyway and can thus safely be changed.