Support dry-run cherry-picks and reverts
This will perform a cherry-pick or revert with the specified arguments,
but not actually create a resulting commit. This allows a user to check
for a `CreateTreeError` on an operation but not modify the repository.

One example use case is in the Delivery team. We want to know if a
security merge request targeting `master` will cherry-pick cleanly to
the previous stable branches. If it does, we can simply cherry-pick it
prior to release; if it doesn't, we need to inform the developer that
they need to create a backport.