Create ansible playbook for demo setup
I was to demo today, and didn't want to spend an hour to configure all
components. Initially the idea was to leverage the Helm charts. Given
the limited experience I had, I wanted to some results to create a set
up for future iterations.

This change is such a first iteration. There's no automation or
integration with the terraform scripts yet.