repository: do not persist config when creating from URL
When creating a new repository from URL, we currently execute the
equivalent of `git clone -c $CFG1`. There's a slight gotcha
here in that `git clone -c` will persist the configuration into the new
repo's gitconfig file, while `git -c $CFG1 clone` does not persist the
configuration to disk. There's two parts we're thus currently persisting
to disk with one being "http.followRedirects" and the other one being
"http.$URL.extraHeader". While the former one doesn't hurt much (but is
not required to be persisted), the extra header is used to pass along
credentials to the remote. As a result, we accidentally persist user
credentials to disk in an unexpected way.

Fix the issue by instead passing all configuration options as global
configuration parameters to `git` instead of passing them to `git
clone`.