rfc: Resource limits through cgroups
To circulate knowledge and organize my own mind, I've written down a
fairly high level document on why cgroups might be a good idea for
Gitaly. Capabilities to limit the blast radius of resource hoarding, be
it intentional or unintential, are rather limited in Gitaly, and we
might need alternatives.