doc: Add an RFC about client authentication and authorization
We currently make strong use of tokens, which are a weak form of
authentication and authorization of clients to access certain
interfaces. But we start to hit some limits as Gitaly deployments are
growing in size with the addition of high availability and strong
consistency, where deployments with at least 6 nodes are the norm.

This commit thus introduces an RFC about improving our security with the
introduction of client-side certificates.