Removal of all http.*.extraHeader config values
Because of incorrect usage of config options passed to clone
command they were stored in repository as a configuration values.
The bad thing that is stores user credentials that could be used
by anyone who has access to configuration of the repository.

With changes made the affected configuration values would be removed
from the repository config when 'OptimizeRepository' rpc will be
called. Eventually it will happen to all repositories and after this
we would be able to remove this code as not needed.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3138 https://gitlab.com/gitlab-org/gitaly/-/issues/2882