Add feature flag for GoUserCherryPick
Prepare the code and the tests for the Go implementation of
UserCherryPick.

The gitaly/service/operations package will test if the feature is
enabled and choose the Ruby or Go implementation based on that. The
latter is still empty for now.

For the tests it involves running the tests twice, once with the
GoUserCherryPick feature flag on, and once off.

Part of https://gitlab.com/gitlab-org/gitaly/-/issues/3071