featureset: Simplify execution with a new `Run()` function
All callsites which use the featureset helper have the same pattern of
creating a feature set, looping around all feature sets, creating a new
testcase, setting up the context and then finally disabling the given
set of features in that context. This is extremely repetitive and easy
to get wrong.

Refactor the code to provide a new `Run()` function which does all of
the test setup and then executes a callback with the modified context as
well as the testcase.