Begin porting tag creation/deletion to Go
Add the bare minimum skeleton for the tag creation/deletion. See[1][2]
for the issues. I'm doing both at the same time because the changes
touch much of the same code, to make merging them down easier.

They'll be protected by flags anyway, so we can turn one or the other
on/off if any bugs are discovered.

See c1e3ccca ("Initial go implementation of UserCreateBranch",
2020-09-30) and c3b32722 ("Initial go implementation of
UserDeleteBranch", 2020-09-29) for some of the prior art I'm ripping
off.

1. https://gitlab.com/gitlab-org/gitaly/-/issues/3064
2. https://gitlab.com/gitlab-org/gitaly/-/issues/3063