Implement basic UserDeleteTagGo
The code is pretty much copy/pasted as-is from c1e3ccca ("Initial go
implementation of UserCreateBranch", 2020-09-30), turns out that doing
this for branches & tags is quite similar.

I'm not doing GetTag() here, which doesn't exist, which would be the
same as the GetBranch(), but GetReference().

TODO: I think the GetBranch() is probably a bug, see GetBranch() in
internal/git/repository.go, i.e. how it DWYM with "refs/heads/foo" and
"foo". Seems like we should do the "refs/heads/" check here earlier
and then use the fully-qualified name in the branch code too.