testing: Add testify/suite to remove boilerplate
This change introduces the suite utilities in one test to showcase how
we can use it to remove boilerplate from our test suite. As first
iteration the setup and teardown of the repository was choosen, though
it can be expanded trivially to include the service server and client.

As with many things, there's a tradeoff when executing tests. For
automated testing `make test` will work as expected, though `go test
-run=` will not work as easily as before. Currently I don't see a way
around that, as the regexp support for the `run` argument requires a
match on the suite name too. `-m` does work.

With the small commit we've got now, upon merging we'd devise a strategy
to include this throughout the Gitaly code base. Though it must be noted
that this can be done piece meal.

Additionally the output changes slightly with this change

```
=== RUN   TestRepositoryService
=== RUN   TestRepositoryService/TestGetInfoAttributesExisting
--- PASS: TestRepositoryService (0.02s)
    --- PASS: TestRepositoryService/TestGetInfoAttributesExisting (0.02s)
    --- PASS: TestRepositoryTestSuite/TestGetInfoAttributesExisting (0.02s)
```