WIP: try to get rid of DecorateError()
I don't like this for the reasons noted in
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2857#note_457436954

Let's try to get rid of it, by making this die very noisily. Then I
can look at the CI fail whale at my leisure and fix any issues if that
seems to make sense, as I did for one file in a commit leading up to
this.

The rest of ErrInvalidArgument() etc. helpers just become a plain
wrapper for the equivalent status/code API use, which (if they don't
fail) we can eventually just use those APIs directly instead.