GoUserCreateTag: add basic skeleton
As in fefc27267 (GoUserDeleteTag: add basic skeleton, 2020-11-16) add
all the non-significant parts of this first. We still call the Ruby
code 100% of the time under the feature. This is just scaffolding for
the test changes.

The lints comments are resolved more generally in [1], but are
explicitly whitelisted for now similar to my 149d56118 (golangci: fix
linting issues on "master", 2020-12-24).

1. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2962