Port UserCreateTag to Go
Add a port of UserCreateTag in Go, fixing [1]. This needed a lot of
test improvements before we could have confidence in this code. See
[2] and [3] for the MRs which did that.

The use of "git mktag" here benefits from my parallel submission to
git.git[4] to make its error messages more sensible. A comment here
lies and says we use "fsck", really in our current version we use
mktag's own fsck-like checks, but that'll be different in future
versions of git. Regardless, as tests in git.git demonstrate in all
the ways that matter for our sanity checking mktag's currently used
code is sufficient to sanity test the input.

See also the draft WIP version of this code in [5] and some comments
by @pks-t there, all of which should be addressed by this version in
[6].

The change to auth_test.go here is something that was preceded by [7],
but which can be fixed now that we're using a Go version of this
code. See the thread at [8].

1. https://gitlab.com/gitlab-org/gitaly/-/issues/3063
2. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2942
3. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2881
4. https://lore.kernel.org/git/20201209200140.29425-1-avarab@gmail.com/
5. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2814
6. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2911
7. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2946
8. https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2911#note_468836913