elect repository specific primary lazily
Electing repository specific primaries eagerly is a lot more work
than electing a primary for a virtual storage. This is due to the
increased number of rows to update. Instead of updating a single
row for a virtual storage, in the worst case Praefect has to update
rows for each repository in a virtual storage.

Currently, excess work is avoided by only running elections when the
health status of the cluster has changed, as opposed to performing
failovers as necessary on every health check like the SQL elector
does. This still requires us to attempt failovers for the whole table
when restarting Praefect due to possibly missing some events while down.
Additionally, if there's a communication error with the database, the
PerRepository elector does not currently re-attempt performing failovers.
This could lead to new primaries not being elected before another health
change occurs in the cluster or a Praefect node is restarted.

To improve the situation further and fix the above problems, this commit
introduces lazy primary election for the PerRepository elector. Rather
than running a background loop, the primary election is performed if
necessary when the primary is needed.

This greatly reduces the amount of work necessary to perform the primary
election as it can happen on per repository basis instead of performing
elections always for every repository.

This also avoid performing elections unnecessarily. If no one tries to
write to a repository while the primary is down, there's no need to
elect a new primary. The primary may come back online before any writes
are attempted and can then continue to service as the primary.

Performing the election as necessary also avoids the problem of failing
to perform failovers and having to wait for another health change event.
Instead, the primary is elected when needed by a request.