combine different read-only metric queries
There's currently two queries for fetching the read-only repository
count from the database, one for virtual storage scoped primaries
and one for repository specific primaries. This makes the maintentance
more difficult as one has to update both queries. Thist commit combines
the queries in to one and makes the primary selecting behavior depend
on the passed in boolean parameter.