account for lazy election when calculating read-only repo count
When virtual storage scoped primaries are in use, there are no lazy elections,
thus any repository with an outdated primary node is in read-only mode.

With repository specific primaries, there's a chance of a lazy fail over
when the primary is needed. Currently the query would consider repositories
which have not yet failed over read-only even if they would fail over and be
writable on the next write request. This commit changes the query to account
for the possiblity of a lazy failover.

To determine whether the repository is in read-only mode, we first check whether
the primary is unhealthy. If the primary is healthy, the elector would not
perform a failover. As such, healthy outdated primaries indicate the repository
is in read-only mode.

If the primary is unhealthy, the elector would attempt to fail over when the
primary is needed. To account for this, the query checks whether there is a
healthy, fully up to date and assigned storage node available. If so, the
elector would be able to fail over to an up to date node and the repository
would not be in read-only mode.