Do not use unsafe "delete" without <oldvalue>
Change the only users of the updateref.Delete() method (aside from its
own tests) to use the safer "Update" instead.

Why not teach Delete() to also take <oldvalue>? The next commit will
do that & change these to use *.Delete() back, but this refactoring
was easier to do atomically this way.