Error helper tests: demonstrate regression in 777a12cfd
The change made in 777a12cfd (wrap errors with grpc context,
2020-05-07) intended to preserve existing GRPC error codes, but as
these tests demonstrate it did the opposite.

We now likely have new code relying on that behavior, so let's
document it in the relevant comments and test for it.

The reason this happens is not because of the removal of the explicit
DecorateError() in that commit, but because of the addition of a
wrapper and the behavior of %v and %w.

The addition of the oldPreconditionFailedf() function demonstrates
that. I.e. the regression didn't have to do with the removal of
DecorateError() (a mere refactoring, we now call
ErrPreconditionFailed() which calls it for us), but the way we're
wrapping the error. See https://blog.golang.org/go1.13-errors and the
%v and %w tests here. I.e. you can still get to the wrapped value, but
what we're going to unpack by default is a different matter.

To test this try:

    # Fails, the old behavior
    git checkout 495a384d4 -- internal/helper/error.go
    go test -v ./internal/helper/ -run TestError -count=1
    # OK, the new (buggy) behavior
    git checkout -f 777a12cfd -- internal/helper/error.go
    go test -v ./internal/helper/ -run TestError -count=1