Error helper tests: refactor recently added tests
Refactor the tests added in the last commit to duplicate fewer test
conditions. We also made the hard assumption that all the
ErrorWhateverf functions did not wrap errors.

Let's add an otherwise unused ErrNotFound() which doesn't have the bug
introduced in 777a12cfd (wrap errors with grpc context, 2020-05-07)
purely to stress the tests themselves in this scenario.

Since I don't like this whole package and intend these test changes as
leading changes to try to get rid of it let's not add it to the main
package and encourage its use, but rather just to the tests. Note that
it uses DecorateError() unlike the non-format functions.