updater: optionally fake the oldvalue for testing
This API allows us to easily inject fake oldvalues into update-ref,
the purpose of which is to easily test ref race conditions without
carefully needing to orchestrate race conditions.

Patrick suggested I could use something like
testhelper.CaptureHookEnv() for this. For update-ref we'd need to have
that more complex since it would need to munge stdin, but more
specifically we'd need to get some out-of-bounds message to it (via a
file) to ask it to map key->value. Not such a big deal now, we could
write this in $PATH:

    #!/bin/sh
    sed 's/EXPECTED/CHANGED/' | git update-ref "$@"

But once we're going to tread update-ref like "cat-file --batch" (via
the transactions it supports) that becomes painful. I.e. a follow-up
to 154dbc1a2 (updateref: Safeguard against accidentally committing
updates, 2020-12-15) with the new interface in Git 2.30.

So I think doing it this way sucks the least, it's obvious from the
variable name that you shouldn't use it for testing, and we could die
on some AreWeTesting() I don't know about in
fakeOldValueForTesting(). The performance hit of checking a hash map
for whether it has any keys is trivial.