operations: Do not run test server in main function
The GitLab server which is being run in the operations' tests main
function doesn't seem to serve any purpose. There is one other test
which does run it manually, and all the others seemingly don't depend on
it being set up.

Remove it and switch `setupAndStartGitlabServer()` to use `testing.TB`
instead of a `testhelper.FatalLogger`. This commit also loses the
explicit sync for the config file as we don't really care whether the
changes are persistent to disk.