UserCreateTag tests: use new update-ref mock to test update-ref error
This fixes a TODO item in be093ba33 (User{Branch,Tag,Submodule}: ferry
update-ref errors upwards, 2020-12-14) for UserDeleteTag, we now know
what response we return for update-ref failures. I'm also adding tests
for the new UserCreateTag.

This is a Go-only test, since in the Ruby codepath we're using
libgit2. I don't think we're going to realistically have both Go &
Ruby tests for features using this deep-gutsy fakery, but at least
here we can see what we'd do in this scenario.