testhelper: split MustRunCommand into that and MayFailRunCommand
Why can't you just use exec.Command? Well, it's annoying to switch
between the two calling conventions, and as a follow-up commit to this
shows it's more obvious what you intend.

I.e. you'd use MayFailRunCommand to cleanup after something that
normally cleans up after itself, but on failure might not (and would
otherwise ruin things for the next test). This is useful for tests of
the form (pseudocode):

    for each test:
        testhelper.MustRunCommand(git tag some-name)
        defer exec.Command(git -d some-name)
        ret = UserDeleteTag(some-name) // maybe this just dies
        [sanity check ret]

In this case usually everything's fine and we don't need the
exec.Command, but if UserDeleteTag() dies we not only fail the test
that died, but also the next test as it expected to be able to setup
"some-name" again.

Yes we could have unique refs or whatever in every test, or we could
eliminate the need to ever use MayFailRunCommand if all "defer" code
did exhaustive "do I need to do this teardown?" checks before doing
anything. But it's more idiomatic and less verbose to just always do
the cleanup unconditionally.