Broke dependency of updateref on config.Config
The updateref package uses a global variable config.Config
that makes testing hard and error prone. It also reduces
flexibility of the configuration of the components.

As we are trying to break dependencies on the global vars
this change removes usage of the config.Config from the
updateref package and substitutes it with passed in config.Cfg
parameter. As package used in different parts of the service
it requires changes in a couple of other places to pass the
parameter from top to bottom.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699