Remove dependency on config.Config from gitaly server factory
As we are trying to remove dependencies on the global vars
this change removes usage of the config.Config from the
GitalyServerFactory and substitutes it with passed in config.Cfg
parameter. As GitalyServerFactory is used in different parts of the
service it requires changes in a couple of other places to pass the
parameter from top to bottom.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2699