handle non-existent start branch in UserCommitFiles
UserCommitFiles receives requests where the StartBranch is set even
for an empty repository. Such requests fail when attempting to resolve
the start branch' commit as the branch does not exist. The Ruby
implementation of the RPC went around this case by checking whether the
repository has no branches. If so, it ignored the start branch. This
scenario caused failures in the QA pipeline. It doesn't appear such
requests are sent in production though.

This commit fixes the failing pipeline by mirroring the Ruby
implementations behavior and ignoring the start branch when the
repository has no branches. The Go implementation already handled this
for the remote repository. This changes the handling to also take place
when the remote repository is not set.