Demo: configuration of direct connection to Postgres
With introduction of database notifications listener we need
to configure direct connection between praefect and postgres
as PgBouncer can't serve LISTEN operation with transaction
scoped configuration.
The change adds praefect instances IPs to the list of allowed IPs
of the Postgres database. So each praefect is allowed to connect
directly to it.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3354