Track feature flags used for RPC call
Feature flags allow us to introduce new functionality with
iterative approach. We can introduce a new feature and protect
its execution with a feature flag, so we could control when
the new code will be executed. It helps a lot in testing on
gitlab.com and reduce time of service unavailability. It is
allowed to enable feature in percents. So we could verify it
on a small subset of all requests. The hard part with it is
to understand if the RPC was called with or without a feature
flag enabled.

The new interceptors for gRPC endpoints extend logging context
with the list of feature flags used during RPC invocation.
As it is now part of the logging context the value will be used
in each log entry. It seems more reasonable then adding additional
log entry as there are not many logs are written by the service
while it operates without issues.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3354