Drop unused assigned column
The column was introduced on the first iteration of adding support
of the variable replication factor, but it turns out it was not an
ideal solution and the repository_assignments table was added instead
to cover missing info. As there is no use of this column we could
drop it with no issues.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3351