golangci: whitelist more common *.Close patterns
Widen the regexp in my 224e7ecb9 (golangci: whitelist common test
patterns, 2020-12-24) to whitelist a few more things, as suggested in
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2962#note_475118299

I'm not just whitelisting any *.Close() as before since that's likely
to catch some false positives. Instead I'm focusing on the common
"defer" patterns.

Before this we'd emit 617 warnings, now we'll emit 603, of those
the *_test.go files went from 425 to 411, so this change only targets
the test suite.