Fix lint issues with c1064ad30
In c1064ad30 (Merge branch 'datastore_create_repo' into 'master',
2021-01-04) I merged a commit which ran afoul of the new checks in
20ef26439 (golangci: Enable error and documentation linters,
2020-12-17).

I think the CI was passing when I looked at the MR and it was based
off an earlier master, but I'm not 100% sure. In any case, this fixes
it.

See also my 149d56118 (golangci: fix linting issues on "master",
2020-12-24) for previous fights with the linter.