testhelper: Convert `NewGitlabTestServer` to receive a testing.TB
The `NewGitlabTestServer` function currently receives a `FatalLogger`
as parameter, which is because the operations tests suite used to set up
a testserver in its main function, which doesn't yet have a `testing.TB`
available. Now that this setup got removed in the preceding commit,
let's convert the code to receive a `testing.TB` instead and use
`require`.