git: Convert subcommands to use a struct
The map of subcommands is currently a simple bitfield used for flags.
While this has served its purpose, it's inflexible and doesn't allow us
to further customize specific commands. While one could do so via flags
as well, it is cumbersome and rather indirect.

Let's refactor the map to use a struct instead. While it currently only
has a `flags` member, it'll get extended in a subsequent commit to also
carry command-specific options.