git: Allow setup of command-specific global options
When executing git-receive-pack or git-upload-pack, we assume that
they'll always get executed with a set of configuration entries which
configure how they behave. It's easy to forget adding those entries
though, and it's not immediately obvious that we do add them in at all
callsites.

In order to improve visibility and guarantee that those options get
injected, this commit extends the git DSL to allow command-specific
global options which always get injected.