Lint: add a "lint-strict" CI target
As noted in
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2962#note_479374433
I think it's more time-efficient to solve most of these issues in
batches. This adds a CI target for that, so we can see what the full
set of things the lint is complaining about is, regardless of how old
they are in the codebase.

We can then use that output to inform edits to .golangci.yml, i.e. to
make certain checks the default.

Or possibly to do choose another approach, such as getting rid of
the--new-from-rev default option in 20ef26439 (golangci: Enable error
and documentation linters, 2020-12-17), add more lints under
"lint-strict", and then only add new default lints once we have them
passing in the codebase in general.