remoterepo: Break upcoming import cycle
We're about to depend on `internal/git/remoterepo` in the repository
service in order to improve the implementation of `FetchSourceBranch`.
Unfortunately, there's currently an import cycle in its tests between
which stops us from doing so.

Move the `repository_test.go` tests into a separate package to break the
import cycle.