repository: Split fetching objects and updating reference
The intention of the `FetchSourceBranch()` RPC is to fetch from an
arbitrary source repository into a target repository. As the source
repository could be hosted on a different Gitaly, we perform a fetch
from source into target repository to obtain missing objects and to
update the reference. While these are two logical steps, those are
currently intermingled as we simply fiddle around with the fetch's
refspec in order to fetch from the source branch into the target branch
directly.

While there's nothing really wrong with this, it does bar us from some
optimizations which we could do if those two steps were logically
separate. Most importantly, if those are separate steps then we can
simply omit the fetch in case source and target repository are the same
as we know that all objects must exist already. So this commit prepares
the code for this optimization by splitting it up into multiple steps:

1. We resolve the object ID of the source branch via an RPC call.
2. We fetch the resolved object ID from the source into the target
   repository.
3. We update the target repository's reference to point to the resolved
   object ID.

So no more fiddling with the refspec.

While at it, this commit also fixes the call to git-fetch(1). We were
previously calling it with the `--prune` flag. This is kind of weird
given that together with the refspec, nothing should've been pruned
anyway. The Ruby implementation doesn't use this flag either, but what
it does add is a `--no-tags` flag. In fact, this flag shouldn't have any
effect either due to the refspec. But let's be consistent here and align
both Go and Ruby implementations by always using `git fetch --no-tags`.