cache: Make locator available to the LeaseKeyer
The cache LeaseKeyer needs to be able to resolve storage locations in
order to identify where temporary data shall be stored. This is
currently done via the global `config.Config`, which we're slowly
getting rid of.

This commit injects a locator into the LeaseKeyer such that we can get
rid of the global config dependency in a subsequent commit.