smarthttp: Get rid of global inforef cache
In order to avoid recomputation of ref advertisements, the smarthttp
service uses a global inforef cache. It being a global variable has the
downside that we cannot inject the Gitaly config into it as it hasn't
yet been initialized when we initialize the cache. But we want to do
that as the cache depends on it to resolve storage locations.

So let's convert the inforef cache from a global variable and move it
into the server struct. This allows us to use the injected locator at a
later point when the cache implementation itself drops the dependency on
the global config.