cache: Introduce a constructor for the LeaseKeyer struct
The LeaseKeyer structure is currently initialized ad-hoc in all
locations. This has been fine for now as the structure hasn't got any
members anyway. But we're about to add a locator to it, and continuing
to use struct initialization directly would not result in any hints by
the compiler that something has changed.

Let's prepare the code for the added member by introducing a constructor
and convert all call sites to use it.