grpc: Convert from `NewServerTLSFromCert` to `NewTLS`
We're currently using `NewServerTLSFromCert()`, which is inflexible and
doesn't allow us to tweak details of the resulting TLS transport
credentials. This commit converts them to use an equivalent invocation
of `NewTLS()` instead, which allows us to specify the `tls.Config`
directly.

No change in behaviour is expected from this commit.