hooks: Drop payload fallback code
In release 13.7, we have migrated hooks away from using multiple
environment variables to instead use a single serialized `HooksPayload`
structure. In order to allow zero-downtime upgrades, we had to keep some
fallback code though to still work with the old environment variables
such that an old server continues to work with a new gitaly-hooks
binary.

Now that 13.7 is out, let's drop the fallback code. We only allow for
zero-downtime upgrades between consecutive releases, so this is safe to
do.