git: Introduce new Revision and ReferenceName types
Our codebase has some confusion with regards to what a revision is and
what a reference is. Furthermore, it's typically not easy to see what
kind of argument a function accepts.

This commit lays the foundation for improving the situation by
introducing two new types Revision and ReferenceName. Subsequent commits
will convert much of our codebase to use these explicit types as
parameters in order to assert that we really only pass expected types to
them.