git: Drop `GetBranch()` function in favor of Reference helpers
The `GetBranch()` function is a small wrapper around `GetReference()`
which simply resolves the given string to a fully qualified reference
first. This commit implements an alternative set of helper functions to
create a Reference from a potentially unqualified branch name and get a
Reference's branch name and converts the codebase to use them instead of
`GetBranch()`.