repository: Silence progress meter of FetchSourceBranch
By default, git-fetch(1) will print the progress of the fetch as well as
all updated references to stderr. Given that we do not catch stderr when
running this command in the FetchSourceBranch RPC, those progress
messages will end up as useless log messages.

Fix the issue by running git-fetch(1) with the "--quiet" flag. This will
cause it to suppress this output while still reporting real errors.