operations: Fix UserFFBranch if called on an ambiguous reference
The UserFFBranch RPC is called with the branch name instead of the fully
qualified name of the branch, e.g. with "master" instead of with
"refs/heads/master". This means that when acting on the branch in the
implementation that we need to make sure that we always convert it to
the fully qualified reference first, otherwise we may easily run into
problems when the name itself is ambiguous. For example if both a tag
and a branch with the same name exists, then we'd resolve to the tag
instead of to the branch.

We're currently not handling this correctly in UserFFBranch. While we do
correctly call `updateReferenceWithHooks` with the fully qualified
reference, we fail to do so when resolving its OID and may thus end up
with the OID of a different reference. Given that we use it to check
whether the merge really is a fast-forward merge, we may thus try to
update a branch even if it's not a fast-forward merge. We're lucky
though: the merge wouldn't ever succeed as `updateReferenceWithHooks`
also receives the wrong old revision -- so it'd only succeed updating
the reference if tag and branch happen to point to the same OID. The
worst that could thus happen is that we simply fail to compute the
fast-forward merge.

Fix the issue by using the fully qualified branch name to resolve the
reference.