objectpool: Count dangling references by going over all refs
We're currently counting references by going over all references in the
"refs/dangling/" namespace. This is perfectly fine, but we're about to
extend counting to "normal" references such that we can compare which
references are more frequent in a repository. In preparation, this
commit thus starts to iterate over all references, only counting those
which start with "refs/dangling".